+++ examples/resources/bunnynet_compute_script_secret/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_compute_script_secret.test "1234|APP_SECRET"