+++ examples/resources/bunnynet_compute_script_secret/resource.tf
@@ -1,0 +1,5 @@
+resource "bunnynet_compute_script_variable" "APP_SECRET" {
+  script = bunnynet_compute_script.test.id
+  name   = "APP_ENV"
+  value  = ""
+}