+++ examples/resources/bunnynet_compute_script_variable/import.sh
@@ -1,1 +1,1 @@
-terraform import bunnynet_compute_script_variable.test "$SCRIPT_ID|$VARIABLE_NAME"
+terraform import bunnynet_compute_script_variable.test "1234|APP_ENV"