+++ internal/provider/provider.go
@@ -104,6 +104,7 @@ func (p *BunnynetProvider) Configure(ctx context.Context, req provider.Configure
 func (p *BunnynetProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewComputeScriptResource,
+		NewComputeScriptSecretResource,
 		NewComputeScriptVariableResource,
 		NewDnsRecordResourceResource,
 		NewDnsZoneResourceResource,